Insert a topic model into a fitted tmfast

Apply varimax rotation for a value of k less than the maximum already included in the tmfast.

Usage

insert_topics(fitted, k, x = NULL)

Arguments

fitted

Fitted tmfast object

k

Desired number of topics for new model

x

Data matrix (document-term matrix), as Matrix object (eg, using build_matrix())

Value

tmfast object, as fitted, with additional topic model inserted

Examples

# \donttest{
set.seed(42)
theta  = rdirichlet(50, 1, k = 4)
phi    = rdirichlet(4, 0.1, k = 20)
corpus = draw_corpus(rep(50L, 50), theta, phi)
model  = tmfast(corpus, n = c(3, 4))
insert_topics(model, k = 2)
#> $totalvar
#> [1] 219.3102
#> 
#> $sdev
#> [1] 9.803849 7.736175 6.104146 2.907624
#> 
#> $rows
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10" "11" "12" "13" "14" "15"
#> [16] "16" "17" "18" "19" "20" "21" "22" "23" "24" "25" "26" "27" "28" "29" "30"
#> [31] "31" "32" "33" "34" "35" "36" "37" "38" "39" "40" "41" "42" "43" "44" "45"
#> [46] "46" "47" "48" "49" "50"
#> 
#> $cols
#>  [1] "3"  "5"  "7"  "8"  "10" "11" "14" "17" "4"  "12" "13" "15" "18" "20" "6" 
#> [16] "19" "2"  "16" "1"  "9" 
#> 
#> $center
#>     3     5     7     8    10    11    14    17     4    12    13    15    18 
#>  5.98  0.76  5.00 10.40  0.50  0.44  9.02  0.90  3.00  2.90  6.20  3.48  0.48 
#>    20     6    19     2    16     1     9 
#>  0.06  0.48  0.18  0.02  0.02  0.16  0.02 
#> 
#> $scale
#> [1] FALSE
#> 
#> $x
#> 50 x 20 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 20 column names ‘3’, ‘5’, ‘7’ ... ]]
#>                                                   
#> 1  18 2  6  1 1 2 17 3  .  .  .  . . . . . . . . .
#> 2   4 2 20  2 . . 11 .  2  1  5  1 2 . . . . . . .
#> 3   . .  6 19 . .  8 1  3  5  4  3 . 1 . . . . . .
#> 4   7 .  2  2 1 1  7 2  5  . 23  . . . . . . . . .
#> 5  15 .  8  6 4 .  8 1  .  .  .  3 . . 4 1 . . . .
#> 6  13 .  2 16 . .  4 1  .  7  2  4 . . 1 . . . . .
#> 7   2 .  8 19 . . 11 1  .  4  .  5 . . . . . . . .
#> 8   2 2  .  8 . 2  2 1  6  4 22  1 . . . . . . . .
#> 9   . .  . 11 . 1  2 1 10  4 15  4 . . 2 . . . . .
#> 10 10 .  .  5 3 .  7 3  5  3 10  3 . 1 . . . . . .
#> 11  7 1  2 11 . .  4 1  2  5  9  7 . . 1 . . . . .
#> 12 12 2  3  5 2 1  8 .  8  .  5  . . . 3 . 1 . . .
#> 13 11 .  1  1 . .  5 2 14  1 11  . . . 3 . . 1 . .
#> 14  8 .  . 15 . 1  2 2  4  5  7  5 . 1 . . . . . .
#> 15 10 3  6  1 1 . 16 1  5  .  6  . . . 1 . . . . .
#> 16  2 .  4 19 . .  2 .  3  2  8 10 . . . . . . . .
#> 17  . 1 16  7 . . 15 .  .  3  4  3 1 . . . . . . .
#> 18  4 .  . 10 . .  5 2  9  3 14  1 . . . . . . 2 .
#> 19  2 .  1 20 1 .  4 .  3  4  6  9 . . . . . . . .
#> 20 13 1  .  8 . 2  7 2  1  5  4  7 . . . . . . . .
#> 21  3 .  . 21 . .  6 .  3  5  1 11 . . . . . . . .
#> 22  . .  3 23 . .  7 .  .  5  . 12 . . . . . . . .
#> 23  2 1  9  6 1 . 12 3  1  4  .  9 2 . . . . . . .
#> 24  1 .  9 15 . . 13 1  .  3  5  1 . . . 2 . . . .
#> 25 11 .  7  4 . 2 14 3  2  1  3  . 1 . 2 . . . . .
#> 26  2 .  . 27 . .  . .  1 11  1  7 . . 1 . . . . .
#> 27 18 3  2  . . 1 18 3  .  2  1  . 1 . 1 . . . . .
#> 28  9 3  2 11 . .  1 1  6  4  9  3 . . . 1 . . . .
#> 29  1 . 10  8 1 . 12 .  4  1 11  . 1 . 1 . . . . .
#> 30  2 .  9  4 . . 20 .  1  . 11  1 1 . . 1 . . . .
#> 31  5 .  4 17 . .  4 .  2  7  7  2 . . . 1 . . 1 .
#> 32  . 1 14 13 1 . 12 .  .  3  .  5 1 . . . . . . .
#> 33 10 2  3  4 1 . 12 .  3  5  6  . 1 . 2 . . . 1 .
#> 34  . .  1 24 . .  4 .  .  8  4  8 . . . . . . 1 .
#> 35 11 1  . 18 1 .  6 1  .  1  3  8 . . . . . . . .
#> 36  4 .  3 13 1 . 14 .  6  .  6  3 . . . . . . . .
#> 37  . 1  . 11 . .  2 4 11  2 16  2 . . . . . . 1 .
#> 38  9 1  5  6 1 1 14 2  .  1  4  6 . . . . . . . .
#> 39  8 . 16  2 . 1 17 .  3  1  1  . . . . 1 . . . .
#> 40  3 2  7 11 . .  6 .  2  6 11  2 . . . . . . . .
#> 41  7 2 20  . . 1 17 .  .  .  .  . 3 . . . . . . .
#> 42 11 1  4 10 . . 12 2  2  3  1  3 1 . . . . . . .
#> 43 10 1  . 16 1 3  5 1  .  2  1  9 . . 1 . . . . .
#> 44  1 1  3  8 . .  5 .  7  1 20  . 2 . . 1 . . . 1
#> 45  . .  1 26 1 .  3 .  1  8  5  4 . . . . . . 1 .
#> 46 11 2  1 14 2 2  4 .  .  1  1  9 1 . 1 1 . . . .
#> 47  . .  1  7 . .  8 . 11  1 19  . 2 . . . . . 1 .
#> 48  . . 14  6 . . 18 .  3  3  1  2 3 . . . . . . .
#> 49  3 1 14  7 . . 20 .  1  .  2  1 1 . . . . . . .
#> 50 17 1  3  2 1 1 20 .  .  .  5  . . . . . . . . .
#> 
#> $rotation
#>                 PC1          PC2          PC3           PC4
#>  [1,] -0.2513465825 -0.006547930  0.774298852 -0.1269740941
#>  [2,] -0.0378986354  0.008060659  0.039992738 -0.0660996799
#>  [3,] -0.3544994814 -0.288926417 -0.525104542 -0.5443809909
#>  [4,]  0.6734086798 -0.361219286 -0.090028626  0.2389383736
#>  [5,] -0.0124105670 -0.003325084  0.053135403 -0.0268283212
#>  [6,] -0.0149682046  0.005973259  0.055678701 -0.0092568100
#>  [7,] -0.5026523359 -0.234671241 -0.157248592  0.7506536627
#>  [8,] -0.0168753798  0.038942595  0.077049040 -0.0007325192
#>  [9,]  0.0370282851  0.366636665 -0.062324223 -0.1024515741
#> [10,]  0.1860045409 -0.079938376 -0.006109357 -0.0840738558
#> [11,]  0.1002361017  0.735928545 -0.243831013  0.1707185567
#> [12,]  0.2347976694 -0.215148389  0.089081846 -0.0521167567
#> [13,] -0.0379864675 -0.008681914 -0.057911433 -0.0391846561
#> [14,]  0.0031559324  0.001580312  0.002814876 -0.0018811767
#> [15,] -0.0133545786  0.016652251  0.060375494 -0.1013682275
#> [16,] -0.0025822997 -0.000943522 -0.006062619 -0.0091009721
#> [17,] -0.0011245687  0.001595050  0.003463324 -0.0057778647
#> [18,] -0.0009406321  0.004778048  0.003012250 -0.0093417319
#> [19,]  0.0114373418  0.014177264 -0.006938100  0.0182940767
#> [20,]  0.0005711820  0.005077513 -0.003344019  0.0009645613
#> 
#> $n
#> [1] 3 4 2
#> 
#> $varimax
#> $varimax$`3`
#> $varimax$`3`$loadings
#>             [,1]          [,2]         [,3]
#> 3   0.0787373504 -1.5005416874  5.114295508
#> 5   0.2195352318 -0.0367052142  0.389861773
#> 7   4.6302283913 -1.6203816519 -1.812253876
#> 8  -5.5354595401 -1.9574186580 -4.150247307
#> 10 -0.0360510410 -0.1201530205  0.323928519
#> 11 -0.0234894272 -0.0572797649  0.367898497
#> 14  4.9015890215 -1.9316433664  0.863067580
#> 17 -0.0754832256  0.1522012054  0.557198275
#> 4  -0.2654361259  2.8579822377  0.288034996
#> 12 -1.5864672421 -0.4230375258 -1.006660070
#> 13 -0.4340360290  5.9483092135 -0.153452241
#> 15 -2.2330130097 -1.5377448837 -1.006891481
#> 18  0.4938375057 -0.0053417665 -0.155759870
#> 20 -0.0358180174  0.0100119423  0.004339573
#> 6  -0.0537564901  0.0160567620  0.407936606
#> 19  0.0395070256  0.0003954456 -0.022422493
#> 2  -0.0001423123  0.0053675113  0.026304265
#> 16 -0.0014899736  0.0299164930  0.029870356
#> 1  -0.0854066540  0.1264243722 -0.055844009
#> 9   0.0026145620  0.0435823562 -0.009204602
#> 
#> $varimax$`3`$rotmat
#>             [,1]        [,2]       [,3]
#> [1,] -0.89259164  0.08867184 -0.4420605
#> [2,] -0.03958277  0.96127223  0.2727433
#> [3,] -0.44912512 -0.26094636  0.8545137
#> 
#> $varimax$`3`$scores
#>            [,1]         [,2]         [,3]
#> 1   0.721185032 -1.002432218  2.010295682
#> 2   1.825165763 -0.023236876 -0.774205532
#> 3  -0.270898947 -0.299878956 -1.261277311
#> 4   0.287010088  2.306231594  0.670794531
#> 5   0.089039945 -0.952253756  1.292423516
#> 6  -1.148678185 -0.887261059  0.757936473
#> 7  -0.073916727 -1.092241346 -1.136986317
#> 8  -0.388787622  2.328674214 -0.123011277
#> 9  -0.602468741  1.688376104 -0.451715261
#> 10 -0.320395484  0.678543362  1.129819192
#> 11 -0.694020720  0.250792666  0.250189067
#> 12 -0.006973304  0.263008055  1.288908981
#> 13 -0.073008708  1.465908156  1.463647432
#> 14 -1.163788423  0.111920767  0.344097206
#> 15  0.973410326  0.172556542  0.972395210
#> 16 -0.949547867  0.057607720 -0.908923879
#> 17  1.543998094 -0.367209955 -1.319116842
#> 18 -0.414209773  1.448287859  0.081430971
#> 19 -1.120667355 -0.177231021 -0.761860801
#> 20 -0.740132899 -0.473078968  1.328473198
#> 21 -1.255098814 -0.861570631 -0.613406163
#> 22 -1.012807829 -1.176721257 -1.294640955
#> 23  0.640488204 -0.745677242 -0.480552101
#> 24  0.510963919 -0.351131776 -1.146828844
#> 25  0.698546154 -0.387530052  0.906269686
#> 26 -1.880405897 -0.905158304 -1.066364003
#> 27  0.526342812 -0.837096923  2.286361684
#> 28 -0.812047095  0.573905134  0.561816944
#> 29  0.985966715  0.755066140 -0.853865578
#> 30  1.557144195  0.496898703 -0.524347044
#> 31 -0.699002380  0.003315443 -0.460539598
#> 32  0.829966263 -0.933337824 -1.447736430
#> 33  0.292244883  0.090817128  0.954982999
#> 34 -1.307342197 -0.594432686 -1.257249580
#> 35 -1.218758364 -0.871600310  0.570397199
#> 36  0.134523314  0.039986754 -0.254542656
#> 37 -0.516012056  1.920848634 -0.408964733
#> 38  0.365244893 -0.519827181  0.620602764
#> 39  1.738036639 -0.599031766 -0.005913658
#> 40  0.057493752  0.596116412 -0.605527133
#> 41  2.203874795 -0.799107855 -0.271825774
#> 42 -0.058512573 -0.797664648  0.732649272
#> 43 -1.202172976 -0.986448864  0.613261852
#> 44  0.128113538  2.135535455 -0.451025492
#> 45 -1.362605662 -0.375132036 -1.340520598
#> 46 -1.087368427 -0.936728828  0.782540859
#> 47  0.236210909  2.252802034 -0.386926869
#> 48  1.652421711 -0.510974960 -1.120430301
#> 49  1.680863320 -0.653558596 -0.787784031
#> 50  0.701373765 -0.519642981  1.896794042
#> 
#> 
#> $varimax$`4`
#> $varimax$`4`$loadings
#>             [,1]         [,2]         [,3]          [,4]
#> 3  -1.322137e+00 -1.768486927  4.534613875 -1.7640384429
#> 5  -1.999171e-01 -0.059106351  0.429655126  0.1020265502
#> 7  -2.339239e+00 -1.633166251 -0.833896593  4.5847238436
#> 8   5.560323e+00 -1.528460903 -3.905928422 -1.9158188522
#> 10 -2.179760e-02 -0.133157943  0.315072799 -0.0963021877
#> 11 -7.469991e-02 -0.075542020  0.328030537 -0.1449681232
#> 14 -5.189748e+00 -2.258563048 -0.485988709  0.9973716671
#> 17 -1.161694e-01  0.123771822  0.483665070 -0.2767502287
#> 4   7.622964e-02  2.855647520  0.477841620 -0.1461351965
#> 12  1.746012e+00 -0.292422386 -0.731166786 -0.3166856536
#> 13 -2.572869e-01  5.934621424 -0.225990832 -0.7097295432
#> 15  2.305209e+00 -1.381664413 -0.801212956 -0.7230980847
#> 18 -2.993548e-01 -0.011595106 -0.080576827  0.4300067201
#> 20  2.986575e-02  0.011388754  0.007187459 -0.0189365606
#> 6   6.391845e-02  0.007582386  0.502231978  0.0072060378
#> 19 -1.250574e-02  0.001162981 -0.006052241  0.0506912462
#> 2   3.066226e-05  0.004584009  0.031328327  0.0007084719
#> 16  3.416982e-03  0.029336943  0.040368599  0.0050285832
#> 1   5.161302e-02  0.130333592 -0.070962030 -0.0673723539
#> 9  -3.762402e-03  0.043735917 -0.008219994  0.0020721071
#> 
#> $varimax$`4`$rotmat
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  0.8547453  0.14851739 -0.3668959 -0.3357685
#> [2,] -0.1169926  0.94458201  0.2594798 -0.1635475
#> [3,]  0.1263950 -0.28983735  0.7307757 -0.6049673
#> [4,] -0.4896401 -0.04125186 -0.5138331 -0.7032257
#> 
#> $varimax$`4`$scores
#>            [,1]        [,2]         [,3]        [,4]
#> 1  -1.344012096 -1.15836442  1.459226758 -0.69966677
#> 2   0.149362264  0.06544791  0.786542772  3.38849596
#> 3   0.435470024 -0.23363487 -1.282367049  0.10495358
#> 4  -0.793331865  2.23353571  0.426378934 -0.46042953
#> 5   0.524124102 -0.94719224  2.048054989  0.86708264
#> 6   1.179746450 -0.84636526  1.069920777 -0.36167305
#> 7   0.205797961 -1.04553471 -1.289816454  0.06584878
#> 8   0.226418981  2.34645978 -0.003861308 -0.19197008
#> 9   0.687796035  1.74710078 -0.149615667  0.04498103
#> 10 -0.112846168  0.62738931  1.004927414 -0.65128755
#> 11  0.798411628  0.29110497  0.570506431 -0.04395984
#> 12  0.001502451  0.22461645  1.535088016  0.03471512
#> 13  0.167432124  1.43751019  1.978061332  0.24640058
#> 14  1.140797762  0.16429449  0.629625149 -0.38155751
#> 15 -1.424288025  0.05170482  0.486047034 -0.31163914
#> 16  1.378795371  0.17400912 -0.416263805  0.30563246
#> 17 -0.569738557 -0.33165373 -0.879419552  1.82866856
#> 18  0.066005965  1.44328472 -0.041054372 -0.52509778
#> 19  1.003731672 -0.10245084 -0.804999937 -0.54591521
#> 20  0.360134499 -0.50559874  1.265752791 -0.78354151
#> 21  0.923763575 -0.80373247 -0.896082821 -0.94134174
#> 22  0.943549525 -1.09043530 -1.501858057 -0.52542142
#> 23 -0.028919646 -0.71705438 -0.127800271  1.01575149
#> 24 -0.531095455 -0.34733511 -1.507238055  0.08253497
#> 25 -0.858357317 -0.46813640  0.714605212 -0.02050076
#> 26  2.101210151 -0.75258909 -0.724577270 -0.35589567
#> 27 -1.704594475 -1.03695246  1.256853318 -1.54090598
#> 28  1.226535870  0.63830588  1.318088663  0.39376973
#> 29 -0.699544945  0.75247766 -0.829134062  0.75552483
#> 30 -2.244359327  0.36999616 -1.606754891 -0.47608168
#> 31  0.945391756  0.07672186 -0.138484235  0.12648293
#> 32  0.182019562 -0.85269140 -0.900638615  1.61711470
#> 33 -0.702615236  0.01139710  0.639672189 -0.47233391
#> 34  1.176713126 -0.49658870 -1.404144962 -0.66349499
#> 35  0.652951795 -0.87139194  0.239293641 -1.24348271
#> 36 -1.001410396 -0.03268278 -1.232187397 -1.20533935
#> 37  0.503591455  1.96587827 -0.196645026 -0.05105479
#> 38 -0.835229295 -0.59981875  0.108918588 -0.58522798
#> 39 -0.919863366 -0.62774125  0.435780564  1.67524536
#> 40  0.446766283  0.65533969 -0.134358629  0.77218820
#> 41 -0.776239097 -0.79474414  0.647764603  2.67646626
#> 42 -0.338071050 -0.85214632  0.376802558 -0.62957862
#> 43  0.922177748 -0.96468353  0.574748274 -0.83509453
#> 44 -0.184357700  2.14305996 -0.402779726  0.10153325
#> 45  1.235460872 -0.27082868 -1.465147298 -0.66188427
#> 46  1.101938406 -0.90149410  1.064010947 -0.36923111
#> 47 -0.761515982  2.21319672 -0.830281414 -0.52876474
#> 48 -1.099094591 -0.52328312 -1.124384427  1.25181001
#> 49 -1.558200841 -0.71417536 -1.205764860  0.63879255
#> 50 -2.199911983 -0.74353146  0.458989204 -1.93162074
#> 
#> 
#> $varimax$`2`
#> $varimax$`2`$loadings
#>            [,1]         [,2]
#> 3  -2.417572287  0.479598519
#> 5  -0.349498525  0.140679267
#> 7  -3.874316434 -1.436852185
#> 8   5.848040051 -4.146779172
#> 10 -0.124356774  0.001000152
#> 11 -0.133402046  0.076639872
#> 14 -5.202792606 -0.715058337
#> 17 -0.096901242  0.329762636
#> 4   0.963540832  2.692274197
#> 12  1.648253930 -0.995525899
#> 13  2.182166495  5.349503831
#> 15  1.890873126 -2.119847023
#> 18 -0.378149122  0.014361355
#> 20  0.032843627  0.005297356
#> 6  -0.100213324  0.153931117
#> 19 -0.026293260 -0.001693399
#> 2  -0.008118586  0.014418973
#> 16 -0.001070363  0.038081747
#> 1   0.133063521  0.083044861
#> 9   0.013902988  0.037162132
#> 
#> $varimax$`2`$rotmat
#>           [,1]       [,2]
#> [1,] 0.9766785 -0.2147069
#> [2,] 0.2147069  0.9766785
#> 
#> $varimax$`2`$scores
#>           [,1]        [,2]
#> 1  -1.67877307 -0.08540979
#> 2  -1.32452724 -0.02186623
#> 3   0.62129948 -0.77297144
#> 4   0.17273390  2.40757140
#> 5  -0.83970674 -0.39532850
#> 6   0.46833247 -0.71800872
#> 7   0.16935814 -1.42675728
#> 8   1.07043777  2.03802808
#> 9   1.19360047  1.29762233
#> 10  0.05920754  0.98331779
#> 11  0.59104107  0.21489985
#> 12 -0.39781073  0.70652967
#> 13 -0.05244779  1.86600561
#> 14  0.92883071  0.04922409
#> 15 -1.16919851  0.65896075
#> 16  1.19194321 -0.42066954
#> 17 -0.97444929 -0.57814695
#> 18  0.75857078  1.29821000
#> 19  1.21888584 -0.60972344
#> 20  0.01713510 -0.06840331
#> 21  1.08131182 -1.20611967
#> 22  1.02992241 -1.70506036
#> 23 -0.60294741 -0.76152130
#> 24 -0.12468191 -0.65915076
#> 25 -1.06666829  0.07784319
#> 26  1.78802209 -1.50552733
#> 27 -1.56184142  0.13660158
#> 28  0.67331040  0.60663721
#> 29 -0.32803448  0.53686241
#> 30 -1.02944237  0.50583186
#> 31  0.78815416 -0.27027905
#> 32 -0.46377913 -1.25458671
#> 33 -0.58703765  0.47327586
#> 34  1.44584721 -1.20101258
#> 35  0.60459873 -0.78208218
#> 36 -0.01170834 -0.03460462
#> 37  1.16965425  1.54012641
#> 38 -0.70542606 -0.19804648
#> 39 -1.70325488 -0.28724007
#> 40  0.35002664  0.33833834
#> 41 -2.07270538 -0.49607942
#> 42 -0.45562341 -0.47795268
#> 43  0.54035221 -0.86969750
#> 44  0.68124640  1.82094447
#> 45  1.58982433 -1.03783647
#> 46  0.39068711 -0.74524094
#> 47  0.59652868  1.96850404
#> 48 -1.18616233 -0.62201229
#> 49 -1.37711576 -0.62862801
#> 50 -1.47752076  0.31462870
#> 
#> 
#> 
#> attr(,"class")
#> [1] "tmfast"    "varimaxes" "list"     
# }