Fit a topic model using PCA+varimax

Fit a topic model using PCA+varimax

Usage

tmfast(dtm, n, row = doc, column = word, value = n, verbose = FALSE, ...)

Arguments

dtm

Document-term matrix. Either an object inheriting from Matrix or a long dataframe representation with row column row, column column column, and value column n.

n

Number of topics to return

row

In dataframe dtm, row column

column

In dataframe dtm, column column

value

In dataframe dtm, value column

verbose

Should irlba() be verbose?

...

Other arguments, passed to varimax_irlba

Value

As per varimax_irlba, of class tmfast

Details

If dtm is not a matrix, will be cast to a sparse matrix using tidytext::case_sparse()